Updated README.md and did some code refactoring
added fake refund function and updated read me
commented out user middleware because it was broken and added comments to every method

+++ app/Models/Ticket/Reservation.php
@@ -8,13 +8,13 @@ use Illuminate\Http\Request;
 
 class Reservation
 {
-    public static function reservation(Request $request) {
+    // add booking to the user account
+    public static function addReservation(Request $request) {
         $paymentType = $request->input('payment_method');
 
         try {
             if ($paymentType) { // Check if payment type is provided
                 $payment = self::getPaymentProcessor($paymentType);
-
                 if ($payment && $payment->processPayment()) { // Process payment and check if successful
                     $T_data = TicketStatus::ACTIVE->value;
                     Ticket::createTicket($request, $T_data);
@@ -29,11 +29,12 @@ class Reservation
 
             \Log::error('Failed to add event: ' . $e->getMessage());
 
-            return redirect()->back()->with('error', 'Failed to add event. Please try again later.');
+            return redirect()->back()->with('error', 'Failed to reserve a ticket. Please try again later.');
         }
 
     }
 
+    // returns with the directory for the class that going to process the payment
     public static function getPaymentProcessor(string $paymentType): ?Payment
     {
         $className = ucfirst(strtolower($paymentType)); // Normalize payment type to class name format (e.g., "debtcard" -> "Debtcard")
@@ -47,27 +48,24 @@ class Reservation
         return null; // Return null if the class doesn't exist or isn't a subclass of Payment
     }
 
+
+
+    //cancel and deletes customer reservation/booking and refunds customer money
     public static function cancelReservation($ticketId)
     {
         try {
-            $ticket = Ticket::where('id', $ticketId)->first() ;
-            $ticket->delete();
-            return redirect()->route('myTickets')->with('success', 'Cancel Ticket Successful');
+            $ticket = Ticket::where('id', $ticketId)->first();
+            $paymentType = $ticket->paymentType;
+            $payment = self::getPaymentProcessor($paymentType);
+            if ($payment && $payment->processPayment()) {
+                $ticket->delete();
+                return redirect()->route('myTickets')->with('success', 'Your booking has been cancelled.');
+            } else {
+                return redirect()->route('myTickets')->with('error', 'Error occurred during the refund process, Please try again.');
+            }
         } catch (\Exception $e) {
-
             \Log::error('Failed to add event: ' . $e->getMessage());
-
-            return redirect()->back()->with('error', 'Failed to add event. Please try again later.');
+            return redirect()->back()->with('error', 'Failed to cancel reservation. Please try again later.');
         }
     }
-
-        //      public  static function validation(Request $request) {
-        ////        dd($request);
-        //        return $request->validate([
-        //            'userId' => 'required|integer|min:5|exists:customers,id',
-        //            'eventId' => 'required|integer|min:5|exists:tickets,id',
-        //            'name' => 'required|string|max:255|exists:customers,name',
-        //            'email' => 'required|email|exists:customers,email'
-        //        ]);
-        //    }
 }
